need to fix the bug

+++ includes/math2.h
@@ -45,7 +45,7 @@ int				quadratic_equation(double *coef, double *answer);
 t_intersect		*init_intersect(void);
 void			find_intersect(t_intersect *is, t_line *l, t_list *obj);
 void			modify_length(double *vector, double length);
-t_line			*new_line(t_camera *camera, double *u, double *v, int *position);
+t_line			*new_line(t_camera *camera, double *u, double *v, double *position);
 double			*cross(double *v1, double *v2);
 double			dot(double *v1, double *v2);
 double			len(double *v);
